Chang long to long long to avoid overflow on 32bits machine

+++ Chapter4/using cast.cpp
@@ -3,7 +3,7 @@
 int main()
 {
     std::cout<<static_cast<double>(6)/5<<std::endl;
-    unsigned long l_value=0x4f4c4c4548;
+    unsigned long long l_value=0x4f4c4c4548;
     // For little end machine, this means HELLO. In ASCII, H is 48, and E is 48, so on.
     auto l_ptr=&l_value;
     char *lchar=reinterpret_cast<char*>(l_ptr);
